requirements.txt
PySide6==6.9.1
base128>=0.1.1
base36>=0.1.1
base58>=2.1.1
base91>=1.0.1
cffi>=1.17.1
chardet>=5.2.0
cryptography>=45.0.3
flowanalyzer>=0.4.0
frida==16.0.8
graphviz>=0.20.3
libnum>=1.7.1
openpyxl>=3.1.5
pandas>=2.2.3
pillow>=11.2.1
pybase100>=0.3.1
pycryptodome>=3.23.0
pyexiv2>=2.15.4
pyopenssl>=25.1.0
pyshark>=0.6
pyside6>=6.9.0
treelib>=1.7.1
yara-python>=4.5.4
binwalk>=2.1.0